requirement.txt
psycopg2-binary==2.9.4
scrapy==2.6.3
scrapy-redis==0.7.3